Add multi-node clustering
Abstract out the mnesia database initialization to a standalone module to make working with a multi-node cluster easier.

Add multi-node clustering for the schema and paste tables. Will backup the existing database contents and restore them when a new node is added.

Add a Vagrant setup where clustering can be tested.